sr: move tests in unit dir
